refactor: 누락된 Address 위치 변경 사항 반영
